Completed deleteMax(), size(lo, hi), getFast(), and putFast(). Started balance()

+++ src/edu/ser222/m03_02/CompletedBST.java
@@ -2,12 +2,13 @@ package edu.ser222.m03_02;
 
 /**
  * A binary search tree based implementation of a symbol table.
- * 
+ * <p>
  * Completion time: (your completion time) TODO: fill in once done
  *
  * @author Tyler Johnson, Sedgewick, Acuna
  * @version 1.0
  */
+
 import java.util.Collections;
 import java.util.LinkedList;
 import java.util.NoSuchElementException;
@@ -32,7 +33,7 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
     public Value get(Key key) {
         Node<Key, Value> iter = root;
 
-        while(iter != null) {
+        while (iter != null) {
             int cmp = key.compareTo(iter.key);
 
             if (cmp < 0)
@@ -79,7 +80,7 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
 
     @Override
     public Key min() {
-        if(root == null)
+        if (root == null)
             throw new NoSuchElementException();
         return min(root).key;
     }
@@ -92,7 +93,7 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
 
     @Override
     public Key max() {
-        if(root == null)
+        if (root == null)
             throw new NoSuchElementException();
         return max(root).key;
     }
@@ -104,7 +105,7 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
 
     @Override
     public Key floor(Key key) {
-        if(root == null)
+        if (root == null)
             throw new NoSuchElementException();
 
         Node<Key, Value> x = floor(root, key);
@@ -133,7 +134,7 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
         if (x == null) return null;
         int t = size(x.left);
         if (t > k) return select(x.left, k);
-        else if (t < k) return select(x.right, k-t-1);
+        else if (t < k) return select(x.right, k - t - 1);
         else return x;
     }
 
@@ -153,7 +154,7 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
 
     @Override
     public void deleteMin() {
-        if(root == null)
+        if (root == null)
             throw new NoSuchElementException();
         root = deleteMin(root);
     }
@@ -175,8 +176,7 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
         int cmp = key.compareTo(x.key);
         if (cmp < 0) x.left = delete(x.left, key);
         else if (cmp > 0) x.right = delete(x.right, key);
-        else
-        {
+        else {
             if (x.right == null) return x.left;
             if (x.left == null) return x.right;
             Node t = x;
@@ -197,15 +197,13 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
     }
 
     @Override
-    public Iterable<Key> keys(Key lo, Key hi)
-    {
+    public Iterable<Key> keys(Key lo, Key hi) {
         Queue<Key> queue = new LinkedList<>();
         keys(root, queue, lo, hi);
         return queue;
     }
 
-    private void keys(Node<Key, Value> x, Queue<Key> queue, Key lo, Key hi)
-    {
+    private void keys(Node<Key, Value> x, Queue<Key> queue, Key lo, Key hi) {
         if (x == null) return;
         int cmplo = lo.compareTo(x.key);
         int cmphi = hi.compareTo(x.key);
@@ -218,6 +216,7 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
         //SKIP, UNNEEDED
         return null;
     }
+
     public Node getRoot() {
         return root;
     }
@@ -230,33 +229,92 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
         return root == null;
     }
 
-    public void deleteMax()  {
+    public void deleteMax() {
+        if (root == null)
+            throw new NoSuchElementException();
         root = deleteMax(root);
     }
 
-    private Node deleteMax(Node x) {
-        if (x.right == null) return x.right;
+    private Node<Key, Value> deleteMax(Node<Key, Value> x) {
+        if (x.right == null) return x.left;
         x.right = deleteMax(x.right);
         x.N = size(x.left) + size(x.right) + 1;
         return x;
     }
 
     public int size(Key lo, Key hi) {
-        // TODO:
-        return 0;
+        Iterable selectedKeys = keys(lo, hi);
+        int count = 0;
+        for (Object elem : selectedKeys) {
+            count++;
+        }
+        return count;
     }
 
     public void putFast(Key key, Value val) {
-        //TODO
+        // TODO: finish this
+        Node<Key, Value> x = root;
+        if (isEmpty()) {
+            root = new Node(key, val, 1);
+            return;
+        }
+        Iterable keys = keys();
+        for (Object elem : keys) {
+            int cmp = key.compareTo((Key) elem);
+            if (cmp == 0) {
+                x.val = val;
+                return;
+            }
+        }
+        while (x != null) {
+            int cmp = key.compareTo(x.key);
+            if (cmp < 0) {
+                if (x.left == null) {
+                    x.N = x.N + 1;
+                    x.left = new Node(key, val, 1);
+                    return;
+                }
+                x.N = x.N + 1;
+                x = x.left;
+            } else if (cmp > 0) {
+                if (x.right == null) {
+                    x.N = x.N + 1;
+                    x.right = new Node(key, val, 1);
+                    return;
+                }
+                x.N = x.N + 1;
+                x = x.right;
+            }
+        }
     }
 
     public Value getFast(Key key) {
-        //TODO
+        Node<Key, Value> x = root;
+        while (x != null) {
+            int cmp = key.compareTo(x.key);
+            if (cmp == 0) return x.val;
+            else if (cmp < 0) x = x.left;
+            else if (cmp > 0) x = x.right;
+        }
         return null;
     }
 
     public void balance() {
         //TODO
+        root = balance(root);
+    }
+
+    private Node<Key, Value> balance(Node<Key, Value> x) {
+        // check if empty?
+        if (x == null) return null;
+        if ((size(x.left) - size(x.right)) > 1) {
+            balance(x.left);
+        } else if ((size(x.right) - size(x.left)) > 1) {
+            balance(x.right);
+        } else {
+            return null; // TODO: fix
+        }
+        return null; // TODO: fix
     }
 
     public String displayLevel(Key key) {
@@ -266,22 +324,38 @@ public class CompletedBST<Key extends Comparable<Key>, Value> implements BST<Key
 
     /**
      * entry point for testing.
-     * 
+     *
      * @param args the command line arguments
      */
     public static void main(String[] args) {
         BST<Integer, String> bst = new CompletedBST();
-        
+
         bst.put(10, "TEN");
         bst.put(3, "THREE");
         bst.put(1, "ONE");
         bst.put(5, "FIVE");
         bst.put(2, "TWO");
         bst.put(7, "SEVEN");
-        
+
+        System.out.println(bst.contains(1));
+        System.out.println(bst.contains(4));
+
+        System.out.println(bst.size(1, 10));
+        System.out.println(bst.size(4, 6));
+
+        System.out.println(bst.getFast(3));
+        System.out.println(bst.getFast(4));
+
+        bst.putFast(4, "FOUR");
+        System.out.println(bst.getFast(4));
+
+        System.out.println("first max: " + bst.max());
+        bst.deleteMax();
+        System.out.println("after delMax: " + bst.max());
+
         System.out.println("Before balance:");
         System.out.println(bst.displayLevel(10)); //root
-        
+
         System.out.println("After balance:");
         bst.balance();
         System.out.println(bst.displayLevel(5)); //root
